requirements.txt
scipy>=1.9.3
numpy>=1.23.4
matplotlib>=3.5.3
torch>=2.0.0+cu118
torchvision>=0.15.1+cu118
tqdm>=4.65.0
Pillow>=8.1.2

opencv-python>=4.6.0.66
torchsummary>=1.5.1
scikit-learn>=1.1.3